+++ terrain/variable_lod/voxel_lod_terrain.cpp
@@ -515,7 +515,7 @@ void VoxelLodTerrain::set_mesh_block_active(VoxelMeshBlockVLT &block, bool activ
 
 // Marks intersecting blocks in the area as modified, updates LODs and schedules remeshing.
 // The provided box must be at LOD0 coordinates.
-void VoxelLodTerrain::post_edit_area(Box3i p_box) {
+void VoxelLodTerrain::post_edit_area(Box3i p_box, bool update_mesh) {
 	ZN_PROFILE_SCOPE();
 	// TODO Better decoupling is needed here.
 	// In the past this padding was necessary for mesh blocks because visuals depend on neighbor voxels.
@@ -527,7 +527,7 @@ void VoxelLodTerrain::post_edit_area(Box3i p_box) {
 	const Box3i box = p_box.padded(1);
 	{
 		MutexLock lock(_update_data->state.blocks_pending_lodding_lod0_mutex);
-		_data->mark_area_modified(box, &_update_data->state.blocks_pending_lodding_lod0);
+		_data->mark_area_modified(box, &_update_data->state.blocks_pending_lodding_lod0, update_mesh);
 	}
 
 #ifdef TOOLS_ENABLED
@@ -536,7 +536,7 @@ void VoxelLodTerrain::post_edit_area(Box3i p_box) {
 	}
 #endif
 
-	if (_instancer != nullptr) {
+	if (_instancer != nullptr && update_mesh) {
 		_instancer->on_area_edited(p_box);
 	}
 }
@@ -1265,7 +1265,12 @@ void VoxelLodTerrain::apply_main_thread_update_tasks() {
 			if (e.tracker->has_next_tasks()) {
 				ERR_PRINT("Completed async edit had next tasks?");
 			}
-			post_edit_area(e.box);
+			post_edit_area(e.box,
+					// Assume the async edit modified voxels in a way it affects the mesh.
+					// Won't be the case if changed only metadata, but so far there is no use case for using an async
+					// edit to change metadata. Metadata is not even used often in smooth terrains (which
+					// VoxelLodTerrain is mostly for)
+					true);
 			return true;
 
 		} else if (e.tracker->is_aborted()) {