+++ doc/source/changelog.md
@@ -48,7 +48,8 @@ Semver is not yet in place, so each version can have breaking changes, although 
         - Fixed graph not always saving when saving the scene
         - Fixed shader generator crash when a node has an unconnected input
     - `VoxelInstanceLibrary`: fixed `find_item_by_name` was not finding items
-    - `VoxelLodTerrain`: fixed error spam when re-generating or destroying the terrain
+    - `VoxelLodTerrain`: Fixed error spam when re-generating or destroying the terrain
+    - `VoxelToolTerrain`: Fixed terrain was not marked as modified when setting voxel metadata
     - `VoxelMeshSDF`: fixed saved resource was not loading properly
 
 - Breaking changes